requirements.txt
# requirements.txt
pymupdf
sentence-transformers
faiss-cpu
matplotlib
gpt4all
numpy
flask
